feat: modularization
This is a rather large commit, but most of the changes are structural.
We introduce a few new targets (and directories in the repository) to
support modularization of the app. In this commit, we're extracting the
`HomeFeature` from the `App` target, encapsulating its state, actions,
and view into a separate target. This means we need a way to transform
the `Store<AppState, AppAction>` to a `Store<HomeState, HomeAction>`,
which is accomplished by the new `Store#view` method.

Most of the other features still need to be extracted, but the process
will be easier now that it's been done once. The steps amount to:

  1. Create targets and associated directories for the feature
  2. Create models for the state and actions of the feature
  3. Remove any dependencies from the feature to the app
  4. Wire the feature into the app using the `Store#view` projection

We'll need some inversion of control code to make this work for the
navigation links in `HomeFeature`, since ideally that feature doesn't
know about the others, even after they've been modularized.